Move the handling of special cases into the emitter
Previously we let the parser handle converting functions like `if()` and `exec2()` into AST nodes that matched their behavior.

After input from @jberg I see that it's cleaner to let the AST always model the syntax (it is called and Abstract _Syntaxt_ Tree after all) and handle the special function types in the emitter.

This simplifies both the parser and the pretty printer.