Support exec2 and exec3This treats exec2/3 as expression blocks (#12)
* Support exec2 and exec3This treats exec2/3 as expression blocks

Each expression is evaluated and the last return value is returned.

This continues the precedence I set with `if()`. Rather than having
a special AST node for these magic functions, I'll use the semantically
correct node (in this case `EXPRESSION_BLOCK`) with a special
`syntax`property to clarify which syntax it used.

This should allow the parser to preserve the syntax used -- which is
important for our pretty printer -- without having this weird syntax
leaking into the emitter/optimizer portions of our code base. This also
reminded me that I should group the `if(a,b,c)` parser as a subtype of
`FUNCTION_CALL` (its syntax) rather than`CONDITIONAL_EXPRESSION` (its
meaning).

* Move the handling of special cases into the emitter

Previously we let the parser handle converting functions like `if()` and `exec2()` into AST nodes that matched their behavior.

After input from @jberg I see that it's cleaner to let the AST always model the syntax (it is called and Abstract _Syntaxt_ Tree after all) and handle the special function types in the emitter.

This simplifies both the parser and the pretty printer.